D Flip-Flop
D flip flop is an sequential circuit which  is known as “delay flip flop” or “data flip flop” which is used to store single bit of data.
The D flip flop has two inputs, data  and clock input which controls the flip flop. when clock input is high, the data is transferred to the output of the flip flop and when the clock input is low, the output of the flip flop is held in its previous state.