Half_Subtractor
Half subtractor is a combination circuit with two inputs and two outputs that are different and borrow. It produces the difference between the two binary bits at the input and also produces an output (Borrow) to indicate if a 1 has been borrowed. In the subtraction (A-B), A is called a Minuend bit and B is called a Subtrahend bit.